+++ src/main/java/kitchenpos/menu/dao/JdbcTemplateMenuProductDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.product.dao;
+package kitchenpos.menu.dao;
 
-import kitchenpos.product.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
